+++ src/plot_generator2d.rs
@@ -16,21 +16,21 @@ pub struct Segment {
 
 // Given the DisplayInfo, it returns an approximation of the plot
 // consistings as a list of segments that should be displayed
-pub fn generate_2dplot(expression: &Expression<f32>, display_info: DisplayInfo) -> Vec<Segment> {
+pub fn generate_2dplot(expression: &Expression<f64>, display_info: DisplayInfo) -> Vec<Segment> {
     let mut segments = Vec::new();
 
     let step = (display_info.x_end - display_info.x_start) / (display_info.resolution as f32);
 
     let mut x_0 = display_info.x_start;
     while x_0 < display_info.x_end {
-        let y_0 = expression.eval_2d(x_0);
+        let y_0 = expression.eval_2d(x_0 as f64);
 
         let x_1 = x_0 + step;
-        let y_1 = expression.eval_2d(x_1);
+        let y_1 = expression.eval_2d(x_1 as f64);
 
         segments.push(Segment {
-            start_point: (x_0, y_0),
-            end_point: (x_1, y_1),
+            start_point: (x_0, y_0 as f32),
+            end_point: (x_1, y_1 as f32),
         });
 
         x_0 += step;
